requirements.txt
pandas
numpy
matplotlib
seaborn
plotly
scikit-learn
statsmodels
prophet
streamlit
yfinance
python-bcb
xgboost
